homework2/pytorch-CycleGAN-and-pix2pix-master/requirements.txt
dominate>=2.4.0
visdom>=0.1.8.9
wandb
pillow>=9.0.0
numpy>=1.21.0
scipy>=1.7.0
requests>=2.27.0
tqdm
scikit-image
matplotlib